DPL Analysis: workaround for arrow not serializing empty tables
