+++ Framework/Core/include/Framework/ASoA.h
@@ -1753,47 +1753,6 @@ auto filter(T&& t, framework::expressions::Filter const& expr)
   return Filtered<T>(t.asArrowTable(), expr);
 }
 
-/// Expression-based column generator to materialize columns
-template <typename... C>
-auto spawner(framework::pack<C...> columns, arrow::Table* atable)
-{
-  static auto new_schema = o2::soa::createSchemaFromColumns(columns);
-  static auto projectors = framework::expressions::createProjectors(columns, atable->schema());
-
-  std::vector<std::shared_ptr<arrow::ChunkedArray>> arrays;
-
-  if (atable->num_rows() == 0) {
-    return arrow::Table::Make(new_schema, arrays);
-  }
-
-  arrow::TableBatchReader reader(*atable);
-  std::shared_ptr<arrow::RecordBatch> batch;
-  arrow::ArrayVector v;
-  std::array<arrow::ArrayVector, sizeof...(C)> chunks;
-
-  while (true) {
-    auto s = reader.ReadNext(&batch);
-    if (!s.ok()) {
-      throw std::runtime_error(fmt::format("Cannot read batches from table: {}", s.ToString()));
-    }
-    if (batch == nullptr) {
-      break;
-    }
-    s = projectors->Evaluate(*batch, arrow::default_memory_pool(), &v);
-    if (!s.ok()) {
-      throw std::runtime_error(fmt::format("Cannot apply projector: {}", s.ToString()));
-    }
-    for (auto i = 0u; i < sizeof...(C); ++i) {
-      chunks[i].emplace_back(v.at(i));
-    }
-  }
-
-  for (auto i = 0u; i < sizeof...(C); ++i) {
-    arrays.push_back(std::make_shared<arrow::ChunkedArray>(chunks[i]));
-  }
-  return arrow::Table::Make(new_schema, arrays);
-}
-
 /// Template for building an index table to access matching rows from non-
 /// joinable, but compatible tables, e.g. Collisions and ZDCs.
 /// First argument is the key table (BCs for the Collisions+ZDCs case), the rest
@@ -1822,26 +1781,6 @@ struct IndexTable : Table<soa::Index<>, H, Ts...> {
 
 template <typename T>
 using is_soa_index_table_t = typename framework::is_base_of_template<soa::IndexTable, T>;
-
-/// On-the-fly adding of expression columns
-template <typename T, typename... Cs>
-auto Extend(T const& table)
-{
-  static_assert((soa::is_type_spawnable_v<Cs> && ...), "You can only extend a table with expression columns");
-  using output_t = Join<T, soa::Table<Cs...>>;
-  return output_t{{spawner(framework::pack<Cs...>{}, table.asArrowTable().get()), table.asArrowTable()}, 0};
-}
-
-/// Template function to attach dynamic columns on-the-fly (e.g. inside
-/// process() function). Dynamic columns need to be compatible with the table.
-template <typename T, typename... Cs>
-auto Attach(T const& table)
-{
-  static_assert((framework::is_base_of_template<o2::soa::DynamicColumn, Cs>::value && ...), "You can only attach dynamic columns");
-  using output_t = Join<T, o2::soa::Table<Cs...>>;
-  return output_t{{table.asArrowTable()}, table.offset()};
-}
-
 } // namespace o2::soa
 
 #endif // O2_FRAMEWORK_ASOA_H_
