+++ Framework/Core/include/Framework/AnalysisHelpers.h
@@ -169,6 +169,7 @@ struct TableTransform {
 template <typename T>
 struct Spawns : TableTransform<typename aod::MetadataTrait<framework::pack_head_t<typename T::originals>>::metadata> {
   using extension_t = framework::pack_head_t<typename T::originals>;
+  using base_table_t = typename aod::MetadataTrait<extension_t>::metadata::base_table_t;
   using expression_pack_t = typename aod::MetadataTrait<extension_t>::metadata::expression_pack_t;
 
   constexpr auto pack()
@@ -540,5 +541,27 @@ struct Partition {
 };
 
 } // namespace o2::framework
+
+namespace o2::soa
+{
+/// On-the-fly adding of expression columns
+template <typename T, typename... Cs>
+auto Extend(T const& table)
+{
+  static_assert((soa::is_type_spawnable_v<Cs> && ...), "You can only extend a table with expression columns");
+  using output_t = Join<T, soa::Table<Cs...>>;
+  return output_t{{o2::framework::spawner(framework::pack<Cs...>{}, table.asArrowTable().get()), table.asArrowTable()}, 0};
+}
+
+/// Template function to attach dynamic columns on-the-fly (e.g. inside
+/// process() function). Dynamic columns need to be compatible with the table.
+template <typename T, typename... Cs>
+auto Attach(T const& table)
+{
+  static_assert((framework::is_base_of_template<o2::soa::DynamicColumn, Cs>::value && ...), "You can only attach dynamic columns");
+  using output_t = Join<T, o2::soa::Table<Cs...>>;
+  return output_t{{table.asArrowTable()}, table.offset()};
+}
+} // namespace o2::soa
 
 #endif // o2_framework_AnalysisHelpers_H_DEFINED
